Renaming to US spelling
